iox-#1614 Move files
Move header file to iox folder, remove 's' from buffer and container
modules
