+++ .clang-tidy-diff-scans.txt
@@ -37,8 +37,9 @@
 ./iceoryx_hoofs/test/moduletests/test_posix*
 ./iceoryx_hoofs/include/iceoryx_hoofs/design_pattern/builder.hpp
 
-./iceoryx_hoofs/containers/include/iox/detail/*
-./iceoryx_hoofs/test/moduletests/test_containers_*
+./iceoryx_hoofs/container/include/iox/*
+./iceoryx_hoofs/container/include/iox/detail/*
+./iceoryx_hoofs/test/moduletests/test_container_*
 
 # IMPORTANT:
 #    after the first # everything is considered a comment, add new files and
