+++ iceoryx_hoofs/CMakeLists.txt
@@ -50,13 +50,13 @@ iox_add_library(
     BUILD_INTERFACE             ${PROJECT_SOURCE_DIR}/include
                                 ${PROJECT_SOURCE_DIR}/legacy/include
                                 ${PROJECT_SOURCE_DIR}/memory/include
-                                ${PROJECT_SOURCE_DIR}/containers/include
+                                ${PROJECT_SOURCE_DIR}/container/include
                                 ${CMAKE_BINARY_DIR}/generated/iceoryx_hoofs/include
     INSTALL_INTERFACE           include/${PREFIX}
     EXPORT_INCLUDE_DIRS         include/
                                 legacy/include/
                                 memory/include/
-                                containers/include/
+                                container/include/
     FILES
         source/concurrent/loffli.cpp
         source/cxx/adaptive_wait.cpp
