+++ iceoryx_hoofs/container/include/iox/detail/uninitialized_array.inl
@@ -15,10 +15,10 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-#ifndef IOX_HOOFS_CONTAINERS_UNINITIALIZED_ARRAY_INL
-#define IOX_HOOFS_CONTAINERS_UNINITIALIZED_ARRAY_INL
+#ifndef IOX_HOOFS_CONTAINER_UNINITIALIZED_ARRAY_INL
+#define IOX_HOOFS_CONTAINER_UNINITIALIZED_ARRAY_INL
 
-#include "iox/detail/uninitialized_array.hpp"
+#include "iox/uninitialized_array.hpp"
 
 namespace iox
 {
@@ -79,4 +79,4 @@ UninitializedArray<ElementType, Capacity, Buffer>::end() const noexcept
 }
 } // namespace iox
 
-#endif // IOX_HOOFS_CONTAINERS_UNINITIALIZED_ARRAY_INL
+#endif // IOX_HOOFS_CONTAINER_UNINITIALIZED_ARRAY_INL
