+++ iceoryx_hoofs/include/iceoryx_hoofs/cxx/vector.hpp
@@ -21,7 +21,7 @@
 #include "iceoryx_hoofs/cxx/attributes.hpp"
 #include "iceoryx_hoofs/cxx/requires.hpp"
 #include "iceoryx_hoofs/log/logging.hpp"
-#include "iox/detail/uninitialized_array.hpp"
+#include "iox/uninitialized_array.hpp"
 
 #include <algorithm>
 #include <cstdint>
