+++ iceoryx_hoofs/include/iceoryx_hoofs/internal/concurrent/fifo.hpp
@@ -18,7 +18,7 @@
 #define IOX_HOOFS_CONCURRENT_FIFO_HPP
 
 #include "iceoryx_hoofs/cxx/optional.hpp"
-#include "iox/detail/uninitialized_array.hpp"
+#include "iox/uninitialized_array.hpp"
 
 #include <atomic>
