+++ iceoryx_hoofs/include/iceoryx_hoofs/posix_wrapper/posix_call.hpp
@@ -20,7 +20,7 @@
 #include "iceoryx_hoofs/cxx/attributes.hpp"
 #include "iceoryx_hoofs/cxx/expected.hpp"
 #include "iceoryx_hoofs/cxx/string.hpp"
-#include "iox/detail/uninitialized_array.hpp"
+#include "iox/uninitialized_array.hpp"
 
 #include <cstdint>
 #include <cstring>
