+++ iceoryx_hoofs/source/posix_wrapper/unix_domain_socket.cpp
@@ -21,7 +21,7 @@
 #include "iceoryx_hoofs/posix_wrapper/posix_call.hpp"
 #include "iceoryx_platform/socket.hpp"
 #include "iceoryx_platform/unistd.hpp"
-#include "iox/detail/uninitialized_array.hpp"
+#include "iox/uninitialized_array.hpp"
 
 #include <chrono>
 #include <string>
